+++ app/src/main/java/com/ibm/rides/appcontroller/RidesApplication.kt
@@ -1,0 +1,7 @@
+package com.ibm.rides.appcontroller
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class RidesApplication : Application()
