+++ app/src/main/java/com/ibm/rides/basecontroller/BaseViewModel.kt
@@ -1,14 +1,13 @@
 package com.ibm.rides.basecontroller
 
-import android.app.Application
 import androidx.databinding.Observable
 import androidx.databinding.PropertyChangeRegistry
-import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.ibm.rides.data.response.Status
 
-abstract class BaseAndroidViewModel(application: Application) : AndroidViewModel(application),
-    Observable {
+abstract class BaseViewModel : ViewModel(), Observable {
+
     private val internetErrorMsg = "Please check your Internet"
     val message = MutableLiveData<String>()
     var networkResponse = MutableLiveData<Status>(null)
@@ -24,7 +23,9 @@ abstract class BaseAndroidViewModel(application: Application) : AndroidViewModel
     }
 
     abstract fun onDestroy()
-    abstract fun getBaseRepository(): BaseRepository?
+
+    /*
+    abstract fun getBaseRepository(): BaseRepository?*/
 
     protected fun showSnackBar(message: String) {
         this.message.postValue(message)
@@ -33,4 +34,9 @@ abstract class BaseAndroidViewModel(application: Application) : AndroidViewModel
     protected fun showInternetError() {
         this.message.postValue(internetErrorMsg)
     }
+}
+
+interface OnApiCallback<T> {
+    fun onSuccess(t: T)
+    fun onFailure(error: String)
 }
