+++ app/src/main/java/com/ibm/rides/data/callback/OnApiCallback.kt
@@ -1,0 +1,6 @@
+package com.ibm.rides.data.callback
+
+interface OnApiCallback<T> {
+    fun onSuccess(t: T)
+    fun onFailure(error: String)
+}
