+++ app/src/main/java/com/ibm/rides/data/response/Status.kt
@@ -1,0 +1,7 @@
+package com.ibm.rides.data.response
+
+enum class Status {
+    SUCCESS,
+    ERROR,
+    LOADING
+}
