--- app/src/main/java/com/ibm/rides/module/vehicle/VehicleActivity.kt
@@ -1,84 +1,0 @@
-package com.ibm.rides.module.vehicle
-
-import android.os.Bundle
-import androidx.navigation.NavController
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.navigateUp
-import androidx.navigation.ui.setupActionBarWithNavController
-import com.ibm.rides.R
-import com.ibm.rides.basecontroller.BaseActivity
-import com.ibm.rides.basecontroller.BaseAndroidViewModel
-import com.ibm.rides.databinding.ActivityVehicleBinding
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
-
-
-class VehicleActivity : BaseActivity() {
-
-    private lateinit var appBarConfiguration: AppBarConfiguration
-    private lateinit var binding: ActivityVehicleBinding
-    private lateinit var navController: NavController
-    private var doubleBackToExitPressedOnce = false
-
-    override fun getViewModel(): BaseAndroidViewModel? {
-        return null
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityVehicleBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        setSupportActionBar(binding.toolbar)
-
-        navController = findNavController(R.id.nav_host_fragment_content_main)
-        appBarConfiguration = AppBarConfiguration(navController.graph)
-        setupActionBarWithNavController(navController, appBarConfiguration)
-
-        /*Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-            .setAction("Action", null).show()*/
-    }
-
-    /*override fun onCreateOptionsMenu(menu: Menu): Boolean {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        menuInflater.inflate(R.menu.menu_main, menu)
-        return true
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        return when (item.itemId) {
-            R.id.action_settings -> true
-            else -> super.onOptionsItemSelected(item)
-        }
-    }*/
-
-    override fun onSupportNavigateUp(): Boolean {
-        val navController = findNavController(R.id.nav_host_fragment_content_main)
-        return navController.navigateUp(appBarConfiguration)
-                || super.onSupportNavigateUp()
-    }
-
-    override fun onBackPressed() {
-        val count = navController.backQueue.count()
-        if (count > 2) {
-            navController.popBackStack()
-        } else {
-            if (doubleBackToExitPressedOnce) {
-                finish()
-                return
-            }
-            this.doubleBackToExitPressedOnce = true
-            showSnackBar("Press again to exit")
-            launch(Dispatchers.Main) {
-                delay(2000)
-                doubleBackToExitPressedOnce = false
-            }
-        }
-    }
-}
