--- app/src/main/java/com/ibm/rides/module/vehicle/userinput/UserInputValidator.kt
@@ -1,10 +1,0 @@
-package com.ibm.rides.module.vehicle.userinput
-
-class UserInputValidator private constructor() {
-
-    companion object {
-        fun isValidVehicleCount(vehicleCount: Int): Boolean {
-            return vehicleCount in 1..100
-        }
-    }
-}
