--- app/src/main/java/com/ibm/rides/network/callback/OnApiCallback.kt
@@ -1,6 +1,0 @@
-package com.ibm.rides.network.callback
-
-interface OnApiCallback<T> {
-    fun onSuccess(t: T)
-    fun onFailure(error: String)
-}
