--- app/src/main/java/com/ibm/rides/network/response/NetworkResponseAdapter.kt
@@ -1,19 +1,0 @@
-package com.ibm.rides.network.response
-
-import okhttp3.ResponseBody
-import retrofit2.Call
-import retrofit2.CallAdapter
-import retrofit2.Converter
-import java.lang.reflect.Type
-
-class NetworkResponseAdapter<S : Any, E : Any>(
-    private val successType: Type,
-    private val errorBodyConverter: Converter<ResponseBody, E>
-) : CallAdapter<S, Call<NetworkResponse<S, E>>> {
-
-    override fun responseType(): Type = successType
-
-    override fun adapt(call: Call<S>): Call<NetworkResponse<S, E>> {
-        return NetworkResponseCall(call, errorBodyConverter)
-    }
-}
