--- app/src/main/java/com/ibm/rides/network/response/Status.kt
@@ -1,7 +1,0 @@
-package com.ibm.rides.network.response
-
-enum class Status {
-    SUCCESS,
-    ERROR,
-    LOADING
-}
