+++ app/src/main/java/com/ibm/rides/ui/vehicle/userinput/UserInputValidator.kt
@@ -1,0 +1,10 @@
+package com.ibm.rides.ui.vehicle.userinput
+
+class UserInputValidator private constructor() {
+
+    companion object {
+        fun isValidVehicleCount(vehicleCount: Int): Boolean {
+            return vehicleCount in 1..100
+        }
+    }
+}
