+++ app/src/main/res/layout/list_item_vehicle.xml
@@ -3,13 +3,6 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
 
-    <data>
-
-        <variable
-            name="viewModel"
-            type="com.ibm.rides.module.vehicle.list.VehicleListingViewModel" />
-    </data>
-
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/containerListItem"
         android:layout_width="match_parent"
